requirements.txt
numpy==1.26.4
pandas>=2.1,<3
google-cloud-bigquery>=3.25.0
db-dtypes>=1.2.0
datasets>=2.20.0
transformers>=4.44.0
accelerate>=0.34.0
bitsandbytes>=0.43.0
peft>=0.12.0
trl>=0.10.1
torch>=2.3.0
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
langchain>=0.2.14
langchain-core>=0.2.34
langchain-community>=0.2.12
langchain-huggingface>=0.0.3
huggingface-hub>=0.24.0
scikit-learn>=1.5.0
python-dotenv>=1.0.1
requests>=2.32.0
wandb>=0.17.0
jupyter>=1.0.0
